Fix tests for 1.6.
contains? is more strict about its arguments, and will throw an
exception if it doesn't support the type of argument you give to it,
so an extra check is added before calling contains? in the catch.

first no longer throws an IllegalArgumentException, so that is
replaced with a call to clojure.string/replace where throwing
IllegalArgumentException is explicit and has been the same since
Clojure 1.2.